+++ MANIFEST.in
@@ -1,8 +1,8 @@
 include README.rst README.dev README.nemo LICENSE HISTORY.txt AUTHORS.txt
 include gsl-config.bat
 include galpy/df/data/*.sav
-include galpy/actionAngle_src/actionAngle_c_ext/*.h
-include galpy/actionAngle_src/actionAngleTorus_c_ext/*.h
+include galpy/actionAngle/actionAngle_c_ext/*.h
+include galpy/actionAngle/actionAngleTorus_c_ext/*.h
 include galpy/orbit_src/orbit_c_ext/*.h
 include galpy/potential_src/potential_c_ext/*.h
 include galpy/util/*.h
