+++ galpy/actionAngle/actionAngleAdiabatic.py
@@ -17,10 +17,10 @@ import numpy as nu
 from galpy.util import galpyWarning
 from galpy.potential import planarPotential, MWPotential
 from galpy.potential_src.Potential import flatten as flatten_potential
-from galpy.actionAngle_src.actionAngleAxi import actionAngleAxi
-from galpy.actionAngle_src.actionAngle import actionAngle
-import galpy.actionAngle_src.actionAngleAdiabatic_c as actionAngleAdiabatic_c
-from galpy.actionAngle_src.actionAngleAdiabatic_c import _ext_loaded as ext_loaded
+from .actionAngleAxi import actionAngleAxi
+from .actionAngle import actionAngle
+from . import actionAngleAdiabatic_c
+from .actionAngleAdiabatic_c import _ext_loaded as ext_loaded
 from galpy.potential_src.Potential import _check_c
 class actionAngleAdiabatic(actionAngle):
     """Action-angle formalism for axisymmetric potentials using the adiabatic approximation"""
