+++ galpy/actionAngle/actionAngleAdiabatic_c.py
