+++ galpy/actionAngle/actionAngleAxi.py
@@ -16,8 +16,8 @@
 import math as m
 import numpy as nu
 from scipy import optimize, integrate
-from galpy.actionAngle_src.actionAngle import *
-from galpy.actionAngle_src.actionAngleVertical import actionAngleVertical
+from .actionAngle import *
+from .actionAngleVertical import actionAngleVertical
 from galpy.potential_src.planarPotential import _evaluateplanarPotentials
 from galpy.potential_src.Potential import epifreq
 from galpy.potential import vcirc
