+++ galpy/actionAngle/actionAngleSpherical.py
@@ -16,8 +16,8 @@ from scipy import integrate
 from galpy.potential import epifreq, omegac
 from galpy.potential_src.Potential import _evaluatePotentials
 from galpy.potential_src.Potential import flatten as flatten_potential
-from galpy.actionAngle_src.actionAngle import *
-from galpy.actionAngle_src.actionAngleAxi import actionAngleAxi, potentialAxi
+from .actionAngle import *
+from .actionAngleAxi import actionAngleAxi, potentialAxi
 class actionAngleSpherical(actionAngle):
     """Action-angle formalism for spherical potentials"""
     def __init__(self,*args,**kwargs):
