+++ setup.py
@@ -143,15 +143,15 @@ orbit_include_dirs= ['galpy/util',
 
 #actionAngleTorus C extension (files here, so we can compile a single extension if so desidered)
 actionAngleTorus_c_src= \
-    glob.glob('galpy/actionAngle_src/actionAngleTorus_c_ext/*.cc')
+    glob.glob('galpy/actionAngle/actionAngleTorus_c_ext/*.cc')
 actionAngleTorus_c_src.extend(\
-    glob.glob('galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/*.cc'))
+    glob.glob('galpy/actionAngle/actionAngleTorus_c_ext/torus/src/*.cc'))
 actionAngleTorus_c_src.extend(\
-    ['galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils/CHB.cc',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils/Err.cc',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils/Compress.cc',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils/Numerics.cc',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils/PJMNum.cc'])
+    ['galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils/CHB.cc',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils/Err.cc',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils/Compress.cc',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils/Numerics.cc',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils/PJMNum.cc'])
 actionAngleTorus_c_src.extend(\
     glob.glob('galpy/potential_src/potential_c_ext/*.c'))
 actionAngleTorus_c_src.extend(\
@@ -160,10 +160,10 @@ actionAngleTorus_c_src.extend(glob.glob('galpy/util/interp_2d/*.c'))
 actionAngleTorus_c_src.extend(glob.glob('galpy/util/*.c'))
 
 actionAngleTorus_include_dirs= \
-    ['galpy/actionAngle_src/actionAngleTorus_c_ext',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src',
-     'galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src/utils',
-     'galpy/actionAngle_src/actionAngle_c_ext',
+    ['galpy/actionAngle/actionAngleTorus_c_ext',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src',
+     'galpy/actionAngle/actionAngleTorus_c_ext/torus/src/utils',
+     'galpy/actionAngle/actionAngle_c_ext',
      'galpy/util/interp_2d',
      'galpy/util',
      'galpy/orbit_src/orbit_c_ext',
@@ -171,9 +171,9 @@ actionAngleTorus_include_dirs= \
 
 if single_ext: #add the code and libraries for the other extensions
     #src
-    orbit_int_c_src.extend(glob.glob('galpy/actionAngle_src/actionAngle_c_ext/*.c'))
+    orbit_int_c_src.extend(glob.glob('galpy/actionAngle/actionAngle_c_ext/*.c'))
     orbit_int_c_src.extend(glob.glob('galpy/potential_src/interppotential_c_ext/*.c'))
-    if os.path.exists('galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src'):
+    if os.path.exists('galpy/actionAngle/actionAngleTorus_c_ext/torus/src'):
         # Add Torus code
         orbit_int_c_src.extend(actionAngleTorus_c_src)
         orbit_int_c_src= list(set(orbit_int_c_src))
@@ -182,14 +182,14 @@ if single_ext: #add the code and libraries for the other extensions
         if not lib in orbit_libraries:
             orbit_libraries.append(lib)
     #includes
-    orbit_include_dirs.extend(['galpy/actionAngle_src/actionAngle_c_ext',
+    orbit_include_dirs.extend(['galpy/actionAngle/actionAngle_c_ext',
                                'galpy/util/interp_2d',
                                'galpy/orbit_src/orbit_c_ext',
                                'galpy/potential_src/potential_c_ext'])
     orbit_include_dirs.extend(['galpy/potential_src/potential_c_ext',
                                'galpy/util/interp_2d',
                                'galpy/util/',
-                               'galpy/actionAngle_src/actionAngle_c_ext',
+                               'galpy/actionAngle/actionAngle_c_ext',
                                'galpy/orbit_src/orbit_c_ext',
                                'galpy/potential_src/interppotential_c_ext'])
     # Add Torus code
@@ -211,12 +211,12 @@ else:
     orbit_int_c_incl= False
 
 #actionAngle C extension
-actionAngle_c_src= glob.glob('galpy/actionAngle_src/actionAngle_c_ext/*.c')
+actionAngle_c_src= glob.glob('galpy/actionAngle/actionAngle_c_ext/*.c')
 actionAngle_c_src.extend(glob.glob('galpy/potential_src/potential_c_ext/*.c'))
 actionAngle_c_src.extend(glob.glob('galpy/util/interp_2d/*.c'))
 actionAngle_c_src.extend(['galpy/util/bovy_symplecticode.c','galpy/util/bovy_rk.c'])
 actionAngle_c_src.append('galpy/orbit_src/orbit_c_ext/integrateFullOrbit.c')
-actionAngle_include_dirs= ['galpy/actionAngle_src/actionAngle_c_ext',
+actionAngle_include_dirs= ['galpy/actionAngle/actionAngle_c_ext',
                            'galpy/orbit_src/orbit_c_ext',
                            'galpy/util/',
                            'galpy/util/interp_2d',
@@ -250,7 +250,7 @@ interppotential_c_src.extend(glob.glob('galpy/util/interp_2d/*.c'))
 interppotential_include_dirs= ['galpy/potential_src/potential_c_ext',
                                'galpy/util/interp_2d',
                                'galpy/util/',
-                               'galpy/actionAngle_src/actionAngle_c_ext',
+                               'galpy/actionAngle/actionAngle_c_ext',
                                'galpy/orbit_src/orbit_c_ext',
                                'galpy/potential_src/interppotential_c_ext']
 
@@ -281,7 +281,7 @@ actionAngleTorus_c= Extension('galpy_actionAngleTorus_c',
                               extra_link_args=extra_link_args)
 if float(gsl_version[0]) >= 1. \
         and (float(gsl_version[0]) >= 2. or float(gsl_version[1]) >= 14.) and \
-        os.path.exists('galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src') and \
+        os.path.exists('galpy/actionAngle/actionAngleTorus_c_ext/torus/src') and \
         not orbit_ext and not interppotential_ext and not single_ext:
     actionAngleTorus_c_incl= True
     ext_modules.append(actionAngleTorus_c)
@@ -299,7 +299,7 @@ setup(name='galpy',
       package_dir = {'galpy/': ''},
       packages=['galpy','galpy/orbit_src','galpy/potential_src',
                 'galpy/df','galpy/util','galpy/snapshot_src',
-                'galpy/actionAngle_src'],
+                'galpy/actionAngle'],
       package_data={'galpy/df':['data/*.sav'],
                     "": ["README.rst","README.dev","LICENSE","AUTHORS.rst"]},
       include_package_data=True,
@@ -361,6 +361,6 @@ if single_ext:
         os.symlink('galpy_integrate_c%s' % _ext_suffix,
                    'galpy_interppotential_c%s' % _ext_suffix)
     if not os.path.exists('galpy_actionAngleTorus_c%s' % _ext_suffix) \
-            and os.path.exists('galpy/actionAngle_src/actionAngleTorus_c_ext/torus/src'):
+            and os.path.exists('galpy/actionAngle/actionAngleTorus_c_ext/torus/src'):
         os.symlink('galpy_integrate_c%s' % _ext_suffix,
                    'galpy_actionAngleTorus_c%s' % _ext_suffix)
